split main.py into multiple files and bug fixes
