+++ update_maker/__main__.py
@@ -1,6 +1,6 @@
-"""Main for the OreSat Linux updater daemon module."""
+"""Main for the OreSat Update Maker module."""
 
-from update_maker.main import main
+from update_maker.cli import main
 
 if __name__ == "__main__":
     main()
